# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'scaffolding'","description":"'Template Spring Boot'","contact":{"name":"404937-Bazan Cristian","email":"404937@tecnicatura.frc.utn.edu.ar"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'Template Spring Boot'"}],"paths":{"/dummy":{"get":{"tags":["dummy-controller"],"operationId":"getDummyList","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DummyDto"}}}}}},"put":{"tags":["dummy-controller"],"operationId":"updateDummy","parameters":[{"name":"dummyDto","in":"query","required":true,"schema":{"$ref":"#/components/schemas/DummyDto"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DummyDto"}}}}}},"post":{"tags":["dummy-controller"],"operationId":"createDummy","parameters":[{"name":"dummyDto","in":"query","required":true,"schema":{"$ref":"#/components/schemas/DummyDto"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DummyDto"}}}}}},"delete":{"tags":["dummy-controller"],"operationId":"deleteDummy","parameters":[{"name":"dummyDto","in":"query","required":true,"schema":{"$ref":"#/components/schemas/DummyDto"}}],"responses":{"200":{"description":"OK"}}}},"/players":{"post":{"tags":["player-controller"],"summary":"Create a new player","description":"Return the player created with your id. If a player exist with the username or email, then return 404. Additionally, the email must be valid and the password must have at least 8 characters and contain at least one number, one lowercase letter, one uppercase letter and one special character","operationId":"savePlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Player"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"400":{"description":"Username or email already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/login":{"post":{"tags":["login-controller"],"summary":"Login a player in the platform","description":"Return the player logged in if the credentials are correct.","operationId":"loginPlayer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Credential"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"404":{"description":"The credentials are invalid","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlayerController"}}}}}}},"/login/v2":{"post":{"tags":["login-controller"],"summary":"Login a player in the platform","description":"Return the player logged in if the credentials are correct","operationId":"loginPlayer_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CredentialV2"}}},"required":true},"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"400":{"description":"Username or email already exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/players/{id}":{"get":{"tags":["player-controller"],"summary":"Get a player by id","description":"Return a player by they id. If the player doesn`t exist return 404","operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Player"}}}},"404":{"description":"Not found","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it`s alive response pong","operationId":"pong","responses":{"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}},"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/dummy/{id}":{"get":{"tags":["dummy-controller"],"operationId":"getDummyList_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DummyDto"}}}}}}}},"components":{"schemas":{"DummyDto":{"type":"object"},"Player":{"required":["email","password","userName"],"type":"object","properties":{"id":{"title":"Player ID","type":"integer","description":"The Player id","format":"int64","example":1},"userName":{"title":"Player user name","type":"string","description":"The Player user name","example":"myUserName"},"password":{"title":"Player password","type":"string","description":"The player password","example":"Password03#"},"email":{"title":"Player email","type":"string","description":"The Player email","example":"email@email.com"},"avatar":{"title":"Player avatar url","type":"string","description":"The Player avatar","nullable":true,"example":"https://localhost:8080/avatars/myUserName"},"lastLoginDate":{"title":"Player last login","type":"string","description":"The Player last login","format":"date-time","nullable":true}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"Credential":{"type":"object","properties":{"identity":{"oneOf":[{"$ref":"#/components/schemas/EmailIdentity"},{"$ref":"#/components/schemas/UserNameIdentity"}]},"password":{"type":"string"}}},"EmailIdentity":{"required":["email","identity_type"],"type":"object","allOf":[{"$ref":"#/components/schemas/Identity"},{"type":"object","properties":{"email":{"title":"Email to legged in","type":"string","description":"The player email","example":"email@email.com"}}}]},"Identity":{"required":["identity_type"],"type":"object","properties":{"identity_type":{"title":"Type of identity to logged in","type":"string","description":"The type of identity provided to logged in","example":"USERNAME or EMAIL","enum":["USERNAME","EMAIL"]}},"discriminator":{"propertyName":"identity_type"}},"UserNameIdentity":{"required":["identity_type","user_name"],"type":"object","allOf":[{"$ref":"#/components/schemas/Identity"},{"type":"object","properties":{"user_name":{"title":"Username","type":"string","description":"The player username","example":"myUserName"}}}]},"PlayerController":{"type":"object"},"CredentialV2":{"required":["identity","password"],"type":"object","properties":{"password":{"type":"string"},"identity":{"title":"Email or UserName to legged in","type":"string","description":"The player email or username","example":"email@email.com or username"}}}}}}